+++ src/lib/map_as3.js
@@ -192,7 +192,7 @@ const createIRule = function createIRule(config) {
             .trim()
             .replace(/\r\n/g, '\n') // unix-style line endings
             .replace(/\s+\n/g, '\n') // trim whitespace on every line
-            .replace(/\\\n/g, '\n') // remove continuation characters
+            .replace(/\\\n[ \t]+/g, '') // remove continuation characters
             .replace(/\n\n/g, '\n \n'); // preserve double newlines
     }
     return { configs: [config] };